refactor(io): extract useProjectLoad from useProjectIO (nua)
Split the 944-line useProjectIO into two single-concern hooks:
- useProjectLoad (574 lines) — owns loadProject and its exclusive params:
  loading overlay, pendingLayoutRefs, updateAppSettings/ProjectSettings,
  all the state setters only called during project open
- useProjectIO (446 lines) — owns save/reload/refresh handlers only

App.tsx calls both hooks at the same call site. No behavioural change.

useProjectLoad doc comment references beads 7c5 (hydrateFromProjectData)
as the long-term fix for the 40-setter coupling that this extraction
intentionally does not address.
